Add a null check if we didn't get a response
If there was any trouble resolving our host information the request would
not be set, and then this would cause null refs deeper in our code.  This
wouldn't be a problem except that our tasks management system is getting
prematurely aborted which causes OnExchangeCodeCompleted never to be
triggered resulting in a stuck Authenication flow.